+++ cookbooks/devel/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file is used to list changes made in each version of devel.
 
+## 0.2.0:
+
+* recipe/default - now includes htop installation and zsh & git recipes
+* recipe/git     - installs and bootstraps git distributed version control system
+* recipe/juju    - now bootstraps juju service orchestration tool
+* recipe/zsh     - installs and bootstraps zsh shell
+
 ## 0.1.0:
 
 * Initial release of devel